@@ -1,20 +1,17 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace Application.DTOs.Account
+namespace Application.DTOs.Account;
+
+public class RefreshToken
 {
-    public class RefreshToken
-    {
-        public int Id { get; set; }
-        public string Token { get; set; }
-        public DateTime Expires { get; set; }
-        public bool IsExpired => DateTime.UtcNow >= Expires;
-        public DateTime Created { get; set; }
-        public string CreatedByIp { get; set; }
-        public DateTime? Revoked { get; set; }
-        public string RevokedByIp { get; set; }
-        public string ReplacedByToken { get; set; }
-        public bool IsActive => Revoked == null && !IsExpired;
-    }
+    public int Id { get; set; }
+    public string Token { get; set; }
+    public DateTime Expires { get; set; }
+    public bool IsExpired => DateTime.UtcNow >= Expires;
+    public DateTime Created { get; set; }
+    public string CreatedByIp { get; set; }
+    public DateTime? Revoked { get; set; }
+    public string RevokedByIp { get; set; }
+    public string ReplacedByToken { get; set; }
+    public bool IsActive => Revoked == null && !IsExpired;
 }
