@@ -1,23 +1,19 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Text;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace Application.DTOs.Account
+namespace Application.DTOs.Account;
+
+public class ResetPasswordRequest
 {
-    public class ResetPasswordRequest
-    {
-        [Required]
-        [EmailAddress]
-        public string Email { get; set; }
-        [Required]
-        public string Token { get; set; }
-        [Required]
-        [MinLength(6)]
-        public string Password { get; set; }
+    [Required]
+    [EmailAddress]
+    public string Email { get; set; }
+    [Required]
+    public string Token { get; set; }
+    [Required]
+    [MinLength(6)]
+    public string Password { get; set; }
 
-        [Required]
-        [Compare("Password")]
-        public string ConfirmPassword { get; set; }
-    }
+    [Required]
+    [Compare("Password")]
+    public string ConfirmPassword { get; set; }
 }
