@@ -1,19 +1,16 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Text;
 
-namespace Application.Exceptions
+namespace Application.Exceptions;
+
+public class ApiException : Exception
 {
-    public class ApiException : Exception
+    public ApiException() : base() { }
+
+    public ApiException(string message) : base(message) { }
+
+    public ApiException(string message, params object[] args)
+        : base(String.Format(CultureInfo.CurrentCulture, message, args))
     {
-        public ApiException() : base() { }
-
-        public ApiException(string message) : base(message) { }
-
-        public ApiException(string message, params object[] args)
-            : base(String.Format(CultureInfo.CurrentCulture, message, args))
-        {
-        }
     }
 }
