@@ -1,26 +1,14 @@
-﻿using FluentValidation.Results;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
-namespace Application.Exceptions
+namespace Application.Exceptions;
+
+public class ValidationException : Exception
 {
-    public class ValidationException : Exception
-    {
-        public ValidationException() : base("One or more validation failures have occurred.")
-        {
-            Errors = new List<string>();
-        }
-        public List<string> Errors { get; }
-        public ValidationException(IEnumerable<ValidationFailure> failures)
-            : this()
-        {
-            foreach (var failure in failures)
-            {
-                Errors.Add(failure.ErrorMessage);
-            }
-        }
-        
-    }
+    public ValidationException() : base("One or more validation failures have occurred.") => Errors = [];
+
+    public ValidationException(IEnumerable<string> errors) : this() => Errors.AddRange(errors);
+
+    public List<string> Errors { get; }
 }
+
