@@ -1,32 +1,34 @@
 ﻿using Application.Exceptions;
 using Application.Interfaces.Repositories;
 using Application.Wrappers;
+
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Text;
+
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Application.Features.Products.Commands.DeleteProductById
+namespace Application.Features.Products.Commands.DeleteProductById;
+
+public class DeleteProductByIdCommand : IRequest<Response<int>>
 {
-    public class DeleteProductByIdCommand : IRequest<Response<int>>
+    public int Id { get; set; }
+    public class DeleteProductByIdCommandHandler : IRequestHandler<DeleteProductByIdCommand, Response<int>>
     {
-        public int Id { get; set; }
-        public class DeleteProductByIdCommandHandler : IRequestHandler<DeleteProductByIdCommand, Response<int>>
+        private readonly IProductRepositoryAsync _productRepository;
+        public DeleteProductByIdCommandHandler(IProductRepositoryAsync productRepository)
         {
-            private readonly IProductRepositoryAsync _productRepository;
-            public DeleteProductByIdCommandHandler(IProductRepositoryAsync productRepository)
+            _productRepository = productRepository;
+        }
+        public async Task<Response<int>> Handle(DeleteProductByIdCommand command, CancellationToken cancellationToken)
+        {
+            Domain.Entities.Product product = await _productRepository.GetByIdAsync(command.Id);
+            if (product == null)
             {
-                _productRepository = productRepository;
+                throw new ApiException($"Product Not Found.");
             }
-            public async Task<Response<int>> Handle(DeleteProductByIdCommand command, CancellationToken cancellationToken)
-            {
-                var product = await _productRepository.GetByIdAsync(command.Id);
-                if (product == null) throw new ApiException($"Product Not Found.");
-                await _productRepository.DeleteAsync(product);
-                return new Response<int>(product.Id);
-            }
+
+            await _productRepository.DeleteAsync(product);
+            return new Response<int>(product.Id);
         }
     }
 }
