@@ -1,12 +1,8 @@
-﻿using Application.Filters;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using Application.Parameters;
 
-namespace Application.Features.Products.Queries.GetAllProducts
+namespace Application.Features.Products.Queries.GetAllProducts;
+
+public class GetAllProductsParameter : RequestParameter
 {
-    public class GetAllProductsParameter : RequestParameter
-    {
 
-    }
 }
