@@ -1,22 +1,19 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Application.Interfaces
+namespace Application.Interfaces;
+
+public interface IGenericRepositoryAsync<T, in TId> where T : class
 {
-    public interface IGenericRepositoryAsync<T,in TId> where T : class
-    {
-        Task<T> GetByIdAsync(TId id);
-        Task<IReadOnlyList<T>> GetAllAsync();
-        Task<IReadOnlyList<T>> GetPagedReponseAsync(int pageNumber, int pageSize);
-        Task<T> AddAsync(T entity);
-        Task UpdateAsync(T entity);
-        Task DeleteAsync(T entity);
-    }
+    Task<T> GetByIdAsync(TId id);
+    Task<IReadOnlyList<T>> GetAllAsync();
+    Task<IReadOnlyList<T>> GetPagedReponseAsync(int pageNumber, int pageSize);
+    Task<T> AddAsync(T entity);
+    Task UpdateAsync(T entity);
+    Task DeleteAsync(T entity);
+}
 
-    public interface IGenericRepositoryAsync<T> : IGenericRepositoryAsync<T, int> where T : class
-    {
+public interface IGenericRepositoryAsync<T> : IGenericRepositoryAsync<T, int> where T : class
+{
 
-    }
 }
