@@ -1,25 +1,25 @@
 ﻿using Application.Behaviours;
-using Application.Features.Products.Commands.CreateProduct;
+
 using AutoMapper;
+
 using FluentValidation;
+
 using MediatR;
+
 using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections.Generic;
+
 using System.Reflection;
-using System.Text;
 
-namespace Application
+namespace Application;
+
+public static class ServiceExtensions
 {
-    public static class ServiceExtensions
+    public static void AddApplicationLayer(this IServiceCollection services)
     {
-        public static void AddApplicationLayer(this IServiceCollection services)
-        {
-            services.AddAutoMapper(Assembly.GetExecutingAssembly());
-            services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
-            services.AddMediatR(Assembly.GetExecutingAssembly());
-            services.AddTransient(typeof(IPipelineBehavior<,>), typeof(ValidationBehavior<,>));
+        _ = services.AddAutoMapper(Assembly.GetExecutingAssembly());
+        _ = services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
+        _ = services.AddMediatR(Assembly.GetExecutingAssembly());
+        _ = services.AddTransient(typeof(IPipelineBehavior<,>), typeof(ValidationBehavior<,>));
 
-        }
     }
 }
