@@ -1,28 +1,25 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 
-namespace Application.Wrappers
+namespace Application.Wrappers;
+
+public class Response<T>
 {
-    public class Response<T>
+    public Response()
     {
-        public Response()
-        {
-        }
-        public Response(T data,string message = null)
-        {
-            Succeeded = true;
-            Message = message;
-            Data = data;
-        }
-        public Response(string message)
-        {
-            Succeeded = false;
-            Message = message;
-        }
-        public bool Succeeded { get; set; }
-        public string Message { get; set; }
-        public List<string> Errors { get; set; }
-        public T Data { get; set; }
     }
+    public Response(T data, string message = null)
+    {
+        Succeeded = true;
+        Message = message;
+        Data = data;
+    }
+    public Response(string message)
+    {
+        Succeeded = false;
+        Message = message;
+    }
+    public bool Succeeded { get; set; }
+    public string Message { get; set; }
+    public List<string> Errors { get; set; }
+    public T Data { get; set; }
 }
