@@ -1,7 +1,6 @@
-﻿namespace Domain.Common
+﻿namespace Domain.Common;
+
+public interface IEntity<T>
 {
-    public interface IEntity<T> 
-    {
-        public T Id { get; set; }
-    }
+    public T Id { get; set; }
 }
