@@ -1,15 +1,11 @@
 ﻿using Domain.Common;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace Domain.Entities
+namespace Domain.Entities;
+
+public class Product : AuditableBaseEntity
 {
-    public class Product : AuditableBaseEntity
-    {
-        public string Name { get; set; }
-        public string Barcode { get; set; }
-        public string Description { get; set; }
-        public decimal Rate { get; set; }
-    }
+    public string Name { get; set; }
+    public string Barcode { get; set; }
+    public string Description { get; set; }
+    public decimal Rate { get; set; }
 }
