@@ -1,37 +1,36 @@
 ﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.IdentityModel.Tokens;
+
 using System;
-using System.Collections.Generic;
 using System.Text;
 
-namespace Infrastructure.Identity.Helpers
+namespace Infrastructure.Identity.Helpers;
+
+public class AuthenticationHelper
 {
-    public class AuthenticationHelper
+    public static void ConfigureService(IServiceCollection service, string Issuer, string Audience, string Key)
     {
-        public static void ConfigureService(IServiceCollection service, string Issuer, string Audience, string Key)
+        _ = service.AddAuthentication(options =>
         {
-            service.AddAuthentication(options =>
+            options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+            options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+        })
+
+            .AddJwtBearer(o =>
             {
-                options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            })
-
-                .AddJwtBearer(o =>
+                o.TokenValidationParameters = new TokenValidationParameters
                 {
-                    o.TokenValidationParameters = new TokenValidationParameters
-                    {
-                        ValidateIssuerSigningKey = true,
-                        ValidateIssuer = true,
-                        ValidateAudience = true,
-                        ValidateLifetime = true,
-                        ClockSkew = TimeSpan.Zero,
+                    ValidateIssuerSigningKey = true,
+                    ValidateIssuer = true,
+                    ValidateAudience = true,
+                    ValidateLifetime = true,
+                    ClockSkew = TimeSpan.Zero,
 
-                        ValidIssuer = Issuer,
-                        ValidAudience = Audience,
-                        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(Key))
-                    };
-                });
-        }
+                    ValidIssuer = Issuer,
+                    ValidAudience = Audience,
+                    IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(Key))
+                };
+            });
     }
 }
