@@ -1,24 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Net;
+﻿using System.Net;
 using System.Net.Sockets;
-using System.Text;
 
-namespace Infrastructure.Identity.Helpers
+namespace Infrastructure.Identity.Helpers;
+
+public class IpHelper
 {
-    public class IpHelper
+    public static string GetIpAddress()
     {
-        public static string GetIpAddress()
+        IPHostEntry host = Dns.GetHostEntry(Dns.GetHostName());
+        foreach (IPAddress ip in host.AddressList)
         {
-            var host = Dns.GetHostEntry(Dns.GetHostName());
-            foreach (var ip in host.AddressList)
+            if (ip.AddressFamily == AddressFamily.InterNetwork)
             {
-                if (ip.AddressFamily == AddressFamily.InterNetwork)
-                {
-                    return ip.ToString();
-                }
+                return ip.ToString();
             }
-            return string.Empty;
         }
+        return string.Empty;
     }
 }
