@@ -1,251 +1,251 @@
-﻿using System;
-using Microsoft.EntityFrameworkCore.Migrations;
+﻿using Microsoft.EntityFrameworkCore.Migrations;
 
-namespace Infrastructure.Identity.Migrations
+using System;
+
+namespace Infrastructure.Identity.Migrations;
+
+public partial class NewSchema : Migration
 {
-    public partial class NewSchema : Migration
+    protected override void Up(MigrationBuilder migrationBuilder)
     {
-        protected override void Up(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.EnsureSchema(
-                name: "Identity");
+        _ = migrationBuilder.EnsureSchema(
+            name: "Identity");
 
-            migrationBuilder.CreateTable(
-                name: "Role",
-                schema: "Identity",
-                columns: table => new
-                {
-                    Id = table.Column<string>(nullable: false),
-                    Name = table.Column<string>(maxLength: 256, nullable: true),
-                    NormalizedName = table.Column<string>(maxLength: 256, nullable: true),
-                    ConcurrencyStamp = table.Column<string>(nullable: true)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_Role", x => x.Id);
-                });
+        _ = migrationBuilder.CreateTable(
+            name: "Role",
+            schema: "Identity",
+            columns: table => new
+            {
+                Id = table.Column<string>(nullable: false),
+                Name = table.Column<string>(maxLength: 256, nullable: true),
+                NormalizedName = table.Column<string>(maxLength: 256, nullable: true),
+                ConcurrencyStamp = table.Column<string>(nullable: true)
+            },
+            constraints: table =>
+            {
+                _ = table.PrimaryKey("PK_Role", x => x.Id);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "User",
-                schema: "Identity",
-                columns: table => new
-                {
-                    Id = table.Column<string>(nullable: false),
-                    UserName = table.Column<string>(maxLength: 256, nullable: true),
-                    NormalizedUserName = table.Column<string>(maxLength: 256, nullable: true),
-                    Email = table.Column<string>(maxLength: 256, nullable: true),
-                    NormalizedEmail = table.Column<string>(maxLength: 256, nullable: true),
-                    EmailConfirmed = table.Column<bool>(nullable: false),
-                    PasswordHash = table.Column<string>(nullable: true),
-                    SecurityStamp = table.Column<string>(nullable: true),
-                    ConcurrencyStamp = table.Column<string>(nullable: true),
-                    PhoneNumber = table.Column<string>(nullable: true),
-                    PhoneNumberConfirmed = table.Column<bool>(nullable: false),
-                    TwoFactorEnabled = table.Column<bool>(nullable: false),
-                    LockoutEnd = table.Column<DateTimeOffset>(nullable: true),
-                    LockoutEnabled = table.Column<bool>(nullable: false),
-                    AccessFailedCount = table.Column<int>(nullable: false),
-                    FirstName = table.Column<string>(nullable: true),
-                    LastName = table.Column<string>(nullable: true)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_User", x => x.Id);
-                });
+        _ = migrationBuilder.CreateTable(
+            name: "User",
+            schema: "Identity",
+            columns: table => new
+            {
+                Id = table.Column<string>(nullable: false),
+                UserName = table.Column<string>(maxLength: 256, nullable: true),
+                NormalizedUserName = table.Column<string>(maxLength: 256, nullable: true),
+                Email = table.Column<string>(maxLength: 256, nullable: true),
+                NormalizedEmail = table.Column<string>(maxLength: 256, nullable: true),
+                EmailConfirmed = table.Column<bool>(nullable: false),
+                PasswordHash = table.Column<string>(nullable: true),
+                SecurityStamp = table.Column<string>(nullable: true),
+                ConcurrencyStamp = table.Column<string>(nullable: true),
+                PhoneNumber = table.Column<string>(nullable: true),
+                PhoneNumberConfirmed = table.Column<bool>(nullable: false),
+                TwoFactorEnabled = table.Column<bool>(nullable: false),
+                LockoutEnd = table.Column<DateTimeOffset>(nullable: true),
+                LockoutEnabled = table.Column<bool>(nullable: false),
+                AccessFailedCount = table.Column<int>(nullable: false),
+                FirstName = table.Column<string>(nullable: true),
+                LastName = table.Column<string>(nullable: true)
+            },
+            constraints: table =>
+            {
+                _ = table.PrimaryKey("PK_User", x => x.Id);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "RoleClaims",
-                schema: "Identity",
-                columns: table => new
-                {
-                    Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:Identity", "1, 1"),
-                    RoleId = table.Column<string>(nullable: false),
-                    ClaimType = table.Column<string>(nullable: true),
-                    ClaimValue = table.Column<string>(nullable: true)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_RoleClaims", x => x.Id);
-                    table.ForeignKey(
-                        name: "FK_RoleClaims_Role_RoleId",
-                        column: x => x.RoleId,
-                        principalSchema: "Identity",
-                        principalTable: "Role",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                });
+        _ = migrationBuilder.CreateTable(
+            name: "RoleClaims",
+            schema: "Identity",
+            columns: table => new
+            {
+                Id = table.Column<int>(nullable: false)
+                    .Annotation("SqlServer:Identity", "1, 1"),
+                RoleId = table.Column<string>(nullable: false),
+                ClaimType = table.Column<string>(nullable: true),
+                ClaimValue = table.Column<string>(nullable: true)
+            },
+            constraints: table =>
+            {
+                _ = table.PrimaryKey("PK_RoleClaims", x => x.Id);
+                _ = table.ForeignKey(
+                    name: "FK_RoleClaims_Role_RoleId",
+                    column: x => x.RoleId,
+                    principalSchema: "Identity",
+                    principalTable: "Role",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "UserClaims",
-                schema: "Identity",
-                columns: table => new
-                {
-                    Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:Identity", "1, 1"),
-                    UserId = table.Column<string>(nullable: false),
-                    ClaimType = table.Column<string>(nullable: true),
-                    ClaimValue = table.Column<string>(nullable: true)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_UserClaims", x => x.Id);
-                    table.ForeignKey(
-                        name: "FK_UserClaims_User_UserId",
-                        column: x => x.UserId,
-                        principalSchema: "Identity",
-                        principalTable: "User",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                });
+        _ = migrationBuilder.CreateTable(
+            name: "UserClaims",
+            schema: "Identity",
+            columns: table => new
+            {
+                Id = table.Column<int>(nullable: false)
+                    .Annotation("SqlServer:Identity", "1, 1"),
+                UserId = table.Column<string>(nullable: false),
+                ClaimType = table.Column<string>(nullable: true),
+                ClaimValue = table.Column<string>(nullable: true)
+            },
+            constraints: table =>
+            {
+                _ = table.PrimaryKey("PK_UserClaims", x => x.Id);
+                _ = table.ForeignKey(
+                    name: "FK_UserClaims_User_UserId",
+                    column: x => x.UserId,
+                    principalSchema: "Identity",
+                    principalTable: "User",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "UserLogins",
-                schema: "Identity",
-                columns: table => new
-                {
-                    LoginProvider = table.Column<string>(nullable: false),
-                    ProviderKey = table.Column<string>(nullable: false),
-                    ProviderDisplayName = table.Column<string>(nullable: true),
-                    UserId = table.Column<string>(nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_UserLogins", x => new { x.LoginProvider, x.ProviderKey });
-                    table.ForeignKey(
-                        name: "FK_UserLogins_User_UserId",
-                        column: x => x.UserId,
-                        principalSchema: "Identity",
-                        principalTable: "User",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                });
+        _ = migrationBuilder.CreateTable(
+            name: "UserLogins",
+            schema: "Identity",
+            columns: table => new
+            {
+                LoginProvider = table.Column<string>(nullable: false),
+                ProviderKey = table.Column<string>(nullable: false),
+                ProviderDisplayName = table.Column<string>(nullable: true),
+                UserId = table.Column<string>(nullable: false)
+            },
+            constraints: table =>
+            {
+                _ = table.PrimaryKey("PK_UserLogins", x => new { x.LoginProvider, x.ProviderKey });
+                _ = table.ForeignKey(
+                    name: "FK_UserLogins_User_UserId",
+                    column: x => x.UserId,
+                    principalSchema: "Identity",
+                    principalTable: "User",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "UserRoles",
-                schema: "Identity",
-                columns: table => new
-                {
-                    UserId = table.Column<string>(nullable: false),
-                    RoleId = table.Column<string>(nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_UserRoles", x => new { x.UserId, x.RoleId });
-                    table.ForeignKey(
-                        name: "FK_UserRoles_Role_RoleId",
-                        column: x => x.RoleId,
-                        principalSchema: "Identity",
-                        principalTable: "Role",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                    table.ForeignKey(
-                        name: "FK_UserRoles_User_UserId",
-                        column: x => x.UserId,
-                        principalSchema: "Identity",
-                        principalTable: "User",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                });
+        _ = migrationBuilder.CreateTable(
+            name: "UserRoles",
+            schema: "Identity",
+            columns: table => new
+            {
+                UserId = table.Column<string>(nullable: false),
+                RoleId = table.Column<string>(nullable: false)
+            },
+            constraints: table =>
+            {
+                _ = table.PrimaryKey("PK_UserRoles", x => new { x.UserId, x.RoleId });
+                _ = table.ForeignKey(
+                    name: "FK_UserRoles_Role_RoleId",
+                    column: x => x.RoleId,
+                    principalSchema: "Identity",
+                    principalTable: "Role",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+                _ = table.ForeignKey(
+                    name: "FK_UserRoles_User_UserId",
+                    column: x => x.UserId,
+                    principalSchema: "Identity",
+                    principalTable: "User",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "UserTokens",
-                schema: "Identity",
-                columns: table => new
-                {
-                    UserId = table.Column<string>(nullable: false),
-                    LoginProvider = table.Column<string>(nullable: false),
-                    Name = table.Column<string>(nullable: false),
-                    Value = table.Column<string>(nullable: true)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_UserTokens", x => new { x.UserId, x.LoginProvider, x.Name });
-                    table.ForeignKey(
-                        name: "FK_UserTokens_User_UserId",
-                        column: x => x.UserId,
-                        principalSchema: "Identity",
-                        principalTable: "User",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                });
+        _ = migrationBuilder.CreateTable(
+            name: "UserTokens",
+            schema: "Identity",
+            columns: table => new
+            {
+                UserId = table.Column<string>(nullable: false),
+                LoginProvider = table.Column<string>(nullable: false),
+                Name = table.Column<string>(nullable: false),
+                Value = table.Column<string>(nullable: true)
+            },
+            constraints: table =>
+            {
+                _ = table.PrimaryKey("PK_UserTokens", x => new { x.UserId, x.LoginProvider, x.Name });
+                _ = table.ForeignKey(
+                    name: "FK_UserTokens_User_UserId",
+                    column: x => x.UserId,
+                    principalSchema: "Identity",
+                    principalTable: "User",
+                    principalColumn: "Id",
+                    onDelete: ReferentialAction.Cascade);
+            });
 
-            migrationBuilder.CreateIndex(
-                name: "RoleNameIndex",
-                schema: "Identity",
-                table: "Role",
-                column: "NormalizedName",
-                unique: true,
-                filter: "[NormalizedName] IS NOT NULL");
+        _ = migrationBuilder.CreateIndex(
+            name: "RoleNameIndex",
+            schema: "Identity",
+            table: "Role",
+            column: "NormalizedName",
+            unique: true,
+            filter: "[NormalizedName] IS NOT NULL");
 
-            migrationBuilder.CreateIndex(
-                name: "IX_RoleClaims_RoleId",
-                schema: "Identity",
-                table: "RoleClaims",
-                column: "RoleId");
+        _ = migrationBuilder.CreateIndex(
+            name: "IX_RoleClaims_RoleId",
+            schema: "Identity",
+            table: "RoleClaims",
+            column: "RoleId");
 
-            migrationBuilder.CreateIndex(
-                name: "EmailIndex",
-                schema: "Identity",
-                table: "User",
-                column: "NormalizedEmail");
+        _ = migrationBuilder.CreateIndex(
+            name: "EmailIndex",
+            schema: "Identity",
+            table: "User",
+            column: "NormalizedEmail");
 
-            migrationBuilder.CreateIndex(
-                name: "UserNameIndex",
-                schema: "Identity",
-                table: "User",
-                column: "NormalizedUserName",
-                unique: true,
-                filter: "[NormalizedUserName] IS NOT NULL");
+        _ = migrationBuilder.CreateIndex(
+            name: "UserNameIndex",
+            schema: "Identity",
+            table: "User",
+            column: "NormalizedUserName",
+            unique: true,
+            filter: "[NormalizedUserName] IS NOT NULL");
 
-            migrationBuilder.CreateIndex(
-                name: "IX_UserClaims_UserId",
-                schema: "Identity",
-                table: "UserClaims",
-                column: "UserId");
+        _ = migrationBuilder.CreateIndex(
+            name: "IX_UserClaims_UserId",
+            schema: "Identity",
+            table: "UserClaims",
+            column: "UserId");
 
-            migrationBuilder.CreateIndex(
-                name: "IX_UserLogins_UserId",
-                schema: "Identity",
-                table: "UserLogins",
-                column: "UserId");
+        _ = migrationBuilder.CreateIndex(
+            name: "IX_UserLogins_UserId",
+            schema: "Identity",
+            table: "UserLogins",
+            column: "UserId");
 
-            migrationBuilder.CreateIndex(
-                name: "IX_UserRoles_RoleId",
-                schema: "Identity",
-                table: "UserRoles",
-                column: "RoleId");
-        }
+        _ = migrationBuilder.CreateIndex(
+            name: "IX_UserRoles_RoleId",
+            schema: "Identity",
+            table: "UserRoles",
+            column: "RoleId");
+    }
 
-        protected override void Down(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.DropTable(
-                name: "RoleClaims",
-                schema: "Identity");
+    protected override void Down(MigrationBuilder migrationBuilder)
+    {
+        _ = migrationBuilder.DropTable(
+            name: "RoleClaims",
+            schema: "Identity");
 
-            migrationBuilder.DropTable(
-                name: "UserClaims",
-                schema: "Identity");
+        _ = migrationBuilder.DropTable(
+            name: "UserClaims",
+            schema: "Identity");
 
-            migrationBuilder.DropTable(
-                name: "UserLogins",
-                schema: "Identity");
+        _ = migrationBuilder.DropTable(
+            name: "UserLogins",
+            schema: "Identity");
 
-            migrationBuilder.DropTable(
-                name: "UserRoles",
-                schema: "Identity");
+        _ = migrationBuilder.DropTable(
+            name: "UserRoles",
+            schema: "Identity");
 
-            migrationBuilder.DropTable(
-                name: "UserTokens",
-                schema: "Identity");
+        _ = migrationBuilder.DropTable(
+            name: "UserTokens",
+            schema: "Identity");
 
-            migrationBuilder.DropTable(
-                name: "Role",
-                schema: "Identity");
+        _ = migrationBuilder.DropTable(
+            name: "Role",
+            schema: "Identity");
 
-            migrationBuilder.DropTable(
-                name: "User",
-                schema: "Identity");
-        }
+        _ = migrationBuilder.DropTable(
+            name: "User",
+            schema: "Identity");
     }
 }
