@@ -1,19 +1,18 @@
 ﻿using Application.DTOs.Account;
-using Microsoft.AspNetCore.Identity;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace Infrastructure.Identity.Models
+using Microsoft.AspNetCore.Identity;
+
+using System.Collections.Generic;
+
+namespace Infrastructure.Identity.Models;
+
+public class ApplicationUser : IdentityUser
 {
-    public class ApplicationUser : IdentityUser
+    public string FirstName { get; set; }
+    public string LastName { get; set; }
+    public List<RefreshToken> RefreshTokens { get; set; }
+    public bool OwnsToken(string token)
     {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public List<RefreshToken> RefreshTokens { get; set; }
-        public bool OwnsToken(string token)
-        {
-            return this.RefreshTokens?.Find(x => x.Token == token) != null;
-        }
+        return this.RefreshTokens?.Find(x => x.Token == token) != null;
     }
 }
