@@ -1,22 +1,21 @@
-﻿using Infrastructure.Identity.Models;
-using Microsoft.AspNetCore.Identity;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using Application.Enums;
+﻿using Application.Enums;
 
-namespace Infrastructure.Identity.Seeds
+using Infrastructure.Identity.Models;
+
+using Microsoft.AspNetCore.Identity;
+
+using System.Threading.Tasks;
+
+namespace Infrastructure.Identity.Seeds;
+
+public static class DefaultRoles
 {
-    public static class DefaultRoles
+    public static async Task SeedAsync(UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
     {
-        public static async Task SeedAsync(UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
-        {
-            //Seed Roles
-            await roleManager.CreateAsync(new IdentityRole(Roles.SuperAdmin.ToString()));
-            await roleManager.CreateAsync(new IdentityRole(Roles.Admin.ToString()));
-            await roleManager.CreateAsync(new IdentityRole(Roles.Moderator.ToString()));
-            await roleManager.CreateAsync(new IdentityRole(Roles.Basic.ToString()));
-        }
+        //Seed Roles
+        _ = await roleManager.CreateAsync(new IdentityRole(Roles.SuperAdmin.ToString()));
+        _ = await roleManager.CreateAsync(new IdentityRole(Roles.Admin.ToString()));
+        _ = await roleManager.CreateAsync(new IdentityRole(Roles.Moderator.ToString()));
+        _ = await roleManager.CreateAsync(new IdentityRole(Roles.Basic.ToString()));
     }
 }
