@@ -1,58 +1,56 @@
 ﻿using Application.Interfaces;
+
 using Domain.Common;
 using Domain.Entities;
+
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
+
 using System.Linq;
-using System.Reflection;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Infrastructure.Persistence.Contexts
+namespace Infrastructure.Persistence.Contexts;
+
+public class ApplicationDbContext : DbContext
 {
-    public class ApplicationDbContext : DbContext
+    private readonly IDateTimeService _dateTime;
+    private readonly IAuthenticatedUserService _authenticatedUser;
+
+    public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options, IDateTimeService dateTime, IAuthenticatedUserService authenticatedUser) : base(options)
     {
-        private readonly IDateTimeService _dateTime;
-        private readonly IAuthenticatedUserService _authenticatedUser;
+        ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
+        _dateTime = dateTime;
+        _authenticatedUser = authenticatedUser;
+    }
+    public DbSet<Product> Products { get; set; }
 
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options, IDateTimeService dateTime, IAuthenticatedUserService authenticatedUser) : base(options)
+    public override Task<int> SaveChangesAsync(CancellationToken cancellationToken = new CancellationToken())
+    {
+        foreach (Microsoft.EntityFrameworkCore.ChangeTracking.EntityEntry<IAuditableEntity> entry in ChangeTracker.Entries<IAuditableEntity>())
         {
-            ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
-            _dateTime = dateTime;
-            _authenticatedUser = authenticatedUser;
-        }
-        public DbSet<Product> Products { get; set; }
-
-        public override Task<int> SaveChangesAsync(CancellationToken cancellationToken = new CancellationToken())
-        {
-            foreach (var entry in ChangeTracker.Entries<IAuditableEntity>())
+            switch (entry.State)
             {
-                switch (entry.State)
-                {
-                    case EntityState.Added:
-                        entry.Entity.Created = _dateTime.NowUtc;
-                        entry.Entity.CreatedBy = _authenticatedUser.UserId;
-                        break;
-                    case EntityState.Modified:
-                        entry.Entity.LastModified = _dateTime.NowUtc;
-                        entry.Entity.LastModifiedBy = _authenticatedUser.UserId;
-                        break;
-                }
+                case EntityState.Added:
+                entry.Entity.Created = _dateTime.NowUtc;
+                entry.Entity.CreatedBy = _authenticatedUser.UserId;
+                break;
+                case EntityState.Modified:
+                entry.Entity.LastModified = _dateTime.NowUtc;
+                entry.Entity.LastModifiedBy = _authenticatedUser.UserId;
+                break;
             }
-            return base.SaveChangesAsync(cancellationToken);
         }
-        protected override void OnModelCreating(ModelBuilder builder)
+        return base.SaveChangesAsync(cancellationToken);
+    }
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+        //All Decimals will have 18,6 Range
+        foreach (Microsoft.EntityFrameworkCore.Metadata.IMutableProperty property in builder.Model.GetEntityTypes()
+        .SelectMany(t => t.GetProperties())
+        .Where(p => p.ClrType == typeof(decimal) || p.ClrType == typeof(decimal?)))
         {
-            //All Decimals will have 18,6 Range
-            foreach (var property in builder.Model.GetEntityTypes()
-            .SelectMany(t => t.GetProperties())
-            .Where(p => p.ClrType == typeof(decimal) || p.ClrType == typeof(decimal?)))
-            {
-                property.SetColumnType("decimal(18,6)");
-            }
-            base.OnModelCreating(builder);
+            property.SetColumnType("decimal(18,6)");
         }
+        base.OnModelCreating(builder);
     }
 }
