@@ -1,37 +1,37 @@
-﻿using System;
-using Microsoft.EntityFrameworkCore.Migrations;
+﻿using Microsoft.EntityFrameworkCore.Migrations;
 
-namespace Infrastructure.Persistence.Migrations
+using System;
+
+namespace Infrastructure.Persistence.Migrations;
+
+public partial class Updat : Migration
 {
-    public partial class Updat : Migration
+    protected override void Up(MigrationBuilder migrationBuilder)
     {
-        protected override void Up(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.CreateTable(
-                name: "Products",
-                columns: table => new
-                {
-                    Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:Identity", "1, 1"),
-                    CreatedBy = table.Column<string>(nullable: true),
-                    Created = table.Column<DateTime>(nullable: false),
-                    LastModifiedBy = table.Column<string>(nullable: true),
-                    LastModified = table.Column<DateTime>(nullable: true),
-                    Name = table.Column<string>(nullable: true),
-                    Barcode = table.Column<string>(nullable: true),
-                    Description = table.Column<string>(nullable: true),
-                    Rate = table.Column<decimal>(type: "decimal(18,6)", nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_Products", x => x.Id);
-                });
-        }
+        _ = migrationBuilder.CreateTable(
+            name: "Products",
+            columns: table => new
+            {
+                Id = table.Column<int>(nullable: false)
+                    .Annotation("SqlServer:Identity", "1, 1"),
+                CreatedBy = table.Column<string>(nullable: true),
+                Created = table.Column<DateTime>(nullable: false),
+                LastModifiedBy = table.Column<string>(nullable: true),
+                LastModified = table.Column<DateTime>(nullable: true),
+                Name = table.Column<string>(nullable: true),
+                Barcode = table.Column<string>(nullable: true),
+                Description = table.Column<string>(nullable: true),
+                Rate = table.Column<decimal>(type: "decimal(18,6)", nullable: false)
+            },
+            constraints: table =>
+            {
+                _ = table.PrimaryKey("PK_Products", x => x.Id);
+            });
+    }
 
-        protected override void Down(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.DropTable(
-                name: "Products");
-        }
+    protected override void Down(MigrationBuilder migrationBuilder)
+    {
+        _ = migrationBuilder.DropTable(
+            name: "Products");
     }
 }
