@@ -1,71 +1,71 @@
 ﻿using Application.Interfaces;
+
 using Infrastructure.Persistence.Contexts;
+
 using Microsoft.EntityFrameworkCore;
-using System;
+
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace Infrastructure.Persistence.Repository
+namespace Infrastructure.Persistence.Repositories;
+
+public class GenericRepositoryAsync<T, TId> : IGenericRepositoryAsync<T, TId> where T : class
 {
-    public class GenericRepositoryAsync<T, TId> : IGenericRepositoryAsync<T,TId> where T : class
+    private readonly ApplicationDbContext _dbContext;
+
+    public GenericRepositoryAsync(ApplicationDbContext dbContext)
     {
-        private readonly ApplicationDbContext _dbContext;
-
-        public GenericRepositoryAsync(ApplicationDbContext dbContext)
-        {
-            _dbContext = dbContext;
-        }
-
-        public virtual async Task<T> GetByIdAsync(TId id)
-        {
-            return await _dbContext.Set<T>().FindAsync(id);
-        }
-
-        public async Task<IReadOnlyList<T>> GetPagedReponseAsync(int pageNumber, int pageSize)
-        {
-            return await _dbContext
-                .Set<T>()
-                .Skip((pageNumber - 1) * pageSize)
-                .Take(pageSize)
-                .AsNoTracking()
-                .ToListAsync();
-        }
-
-        public async Task<T> AddAsync(T entity)
-        {
-            await _dbContext.Set<T>().AddAsync(entity);
-            await _dbContext.SaveChangesAsync();
-            return entity;
-        }
-
-        public async Task UpdateAsync(T entity)
-        {
-            _dbContext.Entry(entity).State = EntityState.Modified;
-            await _dbContext.SaveChangesAsync();
-        }
-
-        public async Task DeleteAsync(T entity)
-        {
-            _dbContext.Set<T>().Remove(entity);
-            await _dbContext.SaveChangesAsync();
-        }
-
-        public async Task<IReadOnlyList<T>> GetAllAsync()
-        {
-            return await _dbContext
-                 .Set<T>()
-                 .ToListAsync();
-        }
-
-       
+        _dbContext = dbContext;
     }
 
-    public class GenericRepositoryAsync<T> : GenericRepositoryAsync<T, int>, IGenericRepositoryAsync<T> where T : class
+    public virtual async Task<T> GetByIdAsync(TId id)
     {
-        public GenericRepositoryAsync(ApplicationDbContext dbContext) : base(dbContext)
-        {
-        }
+        return await _dbContext.Set<T>().FindAsync(id);
+    }
+
+    public async Task<IReadOnlyList<T>> GetPagedReponseAsync(int pageNumber, int pageSize)
+    {
+        return await _dbContext
+            .Set<T>()
+            .Skip((pageNumber - 1) * pageSize)
+            .Take(pageSize)
+            .AsNoTracking()
+            .ToListAsync();
+    }
+
+    public async Task<T> AddAsync(T entity)
+    {
+        _ = await _dbContext.Set<T>().AddAsync(entity);
+        _ = await _dbContext.SaveChangesAsync();
+        return entity;
+    }
+
+    public async Task UpdateAsync(T entity)
+    {
+        _dbContext.Entry(entity).State = EntityState.Modified;
+        _ = await _dbContext.SaveChangesAsync();
+    }
+
+    public async Task DeleteAsync(T entity)
+    {
+        _ = _dbContext.Set<T>().Remove(entity);
+        _ = await _dbContext.SaveChangesAsync();
+    }
+
+    public async Task<IReadOnlyList<T>> GetAllAsync()
+    {
+        return await _dbContext
+             .Set<T>()
+             .ToListAsync();
+    }
+
+
+}
+
+public class GenericRepositoryAsync<T> : GenericRepositoryAsync<T, int>, IGenericRepositoryAsync<T> where T : class
+{
+    public GenericRepositoryAsync(ApplicationDbContext dbContext) : base(dbContext)
+    {
     }
 }
