@@ -1,28 +1,27 @@
 ﻿using Application.Interfaces.Repositories;
+
 using Domain.Entities;
+
 using Infrastructure.Persistence.Contexts;
-using Infrastructure.Persistence.Repository;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
+
 using Microsoft.EntityFrameworkCore;
 
-namespace Infrastructure.Persistence.Repositories
+using System.Threading.Tasks;
+
+namespace Infrastructure.Persistence.Repositories;
+
+public class ProductRepositoryAsync : GenericRepositoryAsync<Product>, IProductRepositoryAsync
 {
-    public class ProductRepositoryAsync : GenericRepositoryAsync<Product>, IProductRepositoryAsync
+    private readonly DbSet<Product> _products;
+
+    public ProductRepositoryAsync(ApplicationDbContext dbContext) : base(dbContext)
     {
-        private readonly DbSet<Product> _products;
+        _products = dbContext.Set<Product>();
+    }
 
-        public ProductRepositoryAsync(ApplicationDbContext dbContext) : base(dbContext)
-        {
-            _products = dbContext.Set<Product>();
-        }
-
-        public Task<bool> IsUniqueBarcodeAsync(string barcode)
-        {
-            return _products
-                .AllAsync(p => p.Barcode != barcode);
-        }
+    public Task<bool> IsUniqueBarcodeAsync(string barcode)
+    {
+        return _products
+            .AllAsync(p => p.Barcode != barcode);
     }
 }
