@@ -1,13 +1,10 @@
 ﻿using Application.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Reflection.Metadata.Ecma335;
-using System.Text;
 
-namespace Infrastructure.Shared.Services
+using System;
+
+namespace Infrastructure.Shared.Services;
+
+public class DateTimeService : IDateTimeService
 {
-    public class DateTimeService : IDateTimeService
-    {
-        public DateTime NowUtc => DateTime.UtcNow;
-    }
+    public DateTime NowUtc => DateTime.UtcNow;
 }
