@@ -1,25 +1,21 @@
 ﻿using Microsoft.AspNetCore.Builder;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+
 using WebApi.Middlewares;
 
-namespace WebApi.Extensions
+namespace WebApi.Extensions;
+
+public static class AppExtensions
 {
-    public static class AppExtensions
+    public static void UseSwaggerExtension(this IApplicationBuilder app)
     {
-        public static void UseSwaggerExtension(this IApplicationBuilder app)
+        _ = app.UseSwagger();
+        _ = app.UseSwaggerUI(c =>
         {
-            app.UseSwagger();
-            app.UseSwaggerUI(c =>
-            {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "CleanArchitecture.WebApi");
-            });
-        }
-        public static void UseErrorHandlingMiddleware(this IApplicationBuilder app)
-        {
-            app.UseMiddleware<ErrorHandlerMiddleware>();
-        }
+            c.SwaggerEndpoint("/swagger/v1/swagger.json", "CleanArchitecture.WebApi");
+        });
+    }
+    public static void UseErrorHandlingMiddleware(this IApplicationBuilder app)
+    {
+        _ = app.UseMiddleware<ErrorHandlerMiddleware>();
     }
 }
