@@ -1,70 +1,68 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.OpenApi.Models;
+
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace WebApi.Extensions
+namespace WebApi.Extensions;
+
+public static class ServiceExtensions
 {
-    public static class ServiceExtensions
+    public static void AddSwaggerExtension(this IServiceCollection services)
     {
-        public static void AddSwaggerExtension(this IServiceCollection services)
+        _ = services.AddSwaggerGen(c =>
         {
-            services.AddSwaggerGen(c =>
+            c.IncludeXmlComments(string.Format(@"{0}\CleanArchitecture.WebApi.xml", System.AppDomain.CurrentDomain.BaseDirectory));
+            c.SwaggerDoc("v1", new OpenApiInfo
             {
-                c.IncludeXmlComments(string.Format(@"{0}\CleanArchitecture.WebApi.xml", System.AppDomain.CurrentDomain.BaseDirectory));
-                c.SwaggerDoc("v1", new OpenApiInfo
+                Version = "v1",
+                Title = "Clean Architecture - WebApi",
+                Description = "This Api will be responsible for overall data distribution and authorization.",
+                Contact = new OpenApiContact
                 {
-                    Version = "v1",
-                    Title = "Clean Architecture - WebApi",
-                    Description = "This Api will be responsible for overall data distribution and authorization.",
-                    Contact = new OpenApiContact
+                    Name = "codewithmukesh",
+                    Email = "hello@codewithmukesh.com",
+                    Url = new Uri("https://codewithmukesh.com/contact"),
+                }
+            });
+            c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
+            {
+                Name = "Authorization",
+                In = ParameterLocation.Header,
+                Type = SecuritySchemeType.ApiKey,
+                Scheme = "Bearer",
+                BearerFormat = "JWT",
+                Description = "Input your Bearer token in this format - Bearer {your token here} to access this API",
+            });
+            c.AddSecurityRequirement(new OpenApiSecurityRequirement
+            {
+                {
+                    new OpenApiSecurityScheme
                     {
-                        Name = "codewithmukesh",
-                        Email = "hello@codewithmukesh.com",
-                        Url = new Uri("https://codewithmukesh.com/contact"),
-                    }
-                });
-                c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
-                {
-                    Name = "Authorization",
-                    In = ParameterLocation.Header,
-                    Type = SecuritySchemeType.ApiKey,
-                    Scheme = "Bearer",
-                    BearerFormat = "JWT",
-                    Description = "Input your Bearer token in this format - Bearer {your token here} to access this API",
-                });
-                c.AddSecurityRequirement(new OpenApiSecurityRequirement
-                {
-                    {
-                        new OpenApiSecurityScheme
+                        Reference = new OpenApiReference
                         {
-                            Reference = new OpenApiReference
-                            {
-                                Type = ReferenceType.SecurityScheme,
-                                Id = "Bearer",
-                            },
-                            Scheme = "Bearer",
-                            Name = "Bearer",
-                            In = ParameterLocation.Header,
-                        }, new List<string>()
-                    },
-                });
+                            Type = ReferenceType.SecurityScheme,
+                            Id = "Bearer",
+                        },
+                        Scheme = "Bearer",
+                        Name = "Bearer",
+                        In = ParameterLocation.Header,
+                    }, new List<string>()
+                },
             });
-        }
-        public static void AddApiVersioningExtension(this IServiceCollection services)
+        });
+    }
+    public static void AddApiVersioningExtension(this IServiceCollection services)
+    {
+        _ = services.AddApiVersioning(config =>
         {
-            services.AddApiVersioning(config =>
-            {
-                // Specify the default API Version as 1.0
-                config.DefaultApiVersion = new ApiVersion(1, 0);
-                // If the client hasn't specified the API version in the request, use the default API version number 
-                config.AssumeDefaultVersionWhenUnspecified = true;
-                // Advertise the API versions supported for the particular endpoint
-                config.ReportApiVersions = true;
-            });
-        }
+            // Specify the default API Version as 1.0
+            config.DefaultApiVersion = new ApiVersion(1, 0);
+            // If the client hasn't specified the API version in the request, use the default API version number 
+            config.AssumeDefaultVersionWhenUnspecified = true;
+            // Advertise the API versions supported for the particular endpoint
+            config.ReportApiVersions = true;
+        });
     }
 }
