@@ -1,60 +1,60 @@
 ﻿using Application.Exceptions;
 using Application.Wrappers;
+
 using Microsoft.AspNetCore.Http;
+
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Net;
 using System.Text.Json;
 using System.Threading.Tasks;
 
-namespace WebApi.Middlewares
+namespace WebApi.Middlewares;
+
+public class ErrorHandlerMiddleware
 {
-    public class ErrorHandlerMiddleware
+    private readonly RequestDelegate _next;
+
+    public ErrorHandlerMiddleware(RequestDelegate next)
     {
-        private readonly RequestDelegate _next;
+        _next = next;
+    }
 
-        public ErrorHandlerMiddleware(RequestDelegate next)
+    public async Task Invoke(HttpContext context)
+    {
+        try
         {
-            _next = next;
+            await _next(context);
         }
-
-        public async Task Invoke(HttpContext context)
+        catch (Exception error)
         {
-            try
-            {
-                await _next(context);
-            }
-            catch (Exception error)
-            {
-                var response = context.Response;
-                response.ContentType = "application/json";
-                var responseModel = new Response<string>() { Succeeded = false, Message = error?.Message };
-                
-                switch (error)
-                {
-                    case Application.Exceptions.ApiException e:
-                        // custom application error
-                        response.StatusCode = (int)HttpStatusCode.BadRequest;
-                        break;
-                    case ValidationException e:
-                        // custom application error
-                        response.StatusCode = (int)HttpStatusCode.BadRequest;
-                        responseModel.Errors = e.Errors;
-                        break;
-                    case KeyNotFoundException e:
-                        // not found error
-                        response.StatusCode = (int)HttpStatusCode.NotFound;
-                        break;
-                    default:
-                        // unhandled error
-                        response.StatusCode = (int)HttpStatusCode.InternalServerError;
-                        break;
-                }
-                var result = JsonSerializer.Serialize(responseModel);
+            HttpResponse response = context.Response;
+            response.ContentType = "application/json";
+            Response<string> responseModel = new() { Succeeded = false, Message = error?.Message };
 
-                await response.WriteAsync(result);
+            switch (error)
+            {
+                case Application.Exceptions.ApiException:
+                // custom application error
+                response.StatusCode = (int)HttpStatusCode.BadRequest;
+                break;
+                case ValidationException e:
+                // custom application error
+                response.StatusCode = (int)HttpStatusCode.BadRequest;
+                responseModel.Errors = e.Errors;
+                break;
+                case KeyNotFoundException:
+                // not found error
+                response.StatusCode = (int)HttpStatusCode.NotFound;
+                break;
+                default:
+                // unhandled error
+                response.StatusCode = (int)HttpStatusCode.InternalServerError;
+                break;
             }
+            string result = JsonSerializer.Serialize(responseModel);
+
+            await response.WriteAsync(result);
         }
     }
 }
