@@ -1,20 +1,17 @@
 ﻿using Application.Interfaces;
+
 using Microsoft.AspNetCore.Http;
-using System;
-using System.Collections.Generic;
-using System.Linq;
+
 using System.Security.Claims;
-using System.Threading.Tasks;
 
-namespace WebApi.Services
+namespace WebApi.Services;
+
+public class AuthenticatedUserService : IAuthenticatedUserService
 {
-    public class AuthenticatedUserService : IAuthenticatedUserService
+    public AuthenticatedUserService(IHttpContextAccessor httpContextAccessor)
     {
-        public AuthenticatedUserService(IHttpContextAccessor httpContextAccessor)
-        {
-            UserId = httpContextAccessor.HttpContext?.User?.FindFirstValue("uid");
-        }
-
-        public string UserId { get; }
+        UserId = httpContextAccessor.HttpContext?.User?.FindFirstValue("uid");
     }
+
+    public string UserId { get; }
 }
