@@ -1,14 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Application.Enums;
 
-namespace Application.Enums
+public enum Roles
 {
-    public enum Roles
-    {
-        SuperAdmin,
-        Admin,
-        Moderator,
-        Basic
-    }
+    SuperAdmin,
+    Admin,
+    Moderator,
+    Basic
 }
