@@ -1,37 +1,36 @@
-﻿using Application.Filters;
-using Application.Interfaces.Repositories;
+﻿using Application.Interfaces.Repositories;
 using Application.Wrappers;
+
 using AutoMapper;
+
 using MediatR;
-using System;
+
 using System.Collections.Generic;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Application.Features.Products.Queries.GetAllProducts
-{
-    public class GetAllProductsQuery : IRequest<PagedResponse<IEnumerable<GetAllProductsViewModel>>>
-    {
-        public int PageNumber { get; set; }
-        public int PageSize { get; set; }
-    }
-    public class GetAllProductsQueryHandler : IRequestHandler<GetAllProductsQuery, PagedResponse<IEnumerable<GetAllProductsViewModel>>>
-    {
-        private readonly IProductRepositoryAsync _productRepository;
-        private readonly IMapper _mapper;
-        public GetAllProductsQueryHandler(IProductRepositoryAsync productRepository, IMapper mapper)
-        {
-            _productRepository = productRepository;
-            _mapper = mapper;
-        }
+namespace Application.Features.Products.Queries.GetAllProducts;
 
-        public async Task<PagedResponse<IEnumerable<GetAllProductsViewModel>>> Handle(GetAllProductsQuery request, CancellationToken cancellationToken)
-        {
-            var validFilter = _mapper.Map<GetAllProductsParameter>(request);
-            var product = await _productRepository.GetPagedReponseAsync(validFilter.PageNumber,validFilter.PageSize);
-            var productViewModel = _mapper.Map<IEnumerable<GetAllProductsViewModel>>(product);
-            return new PagedResponse<IEnumerable<GetAllProductsViewModel>>(productViewModel, validFilter.PageNumber, validFilter.PageSize);           
-        }
+public class GetAllProductsQuery : IRequest<PagedResponse<IEnumerable<GetAllProductsViewModel>>>
+{
+    public int PageNumber { get; set; }
+    public int PageSize { get; set; }
+}
+public class GetAllProductsQueryHandler : IRequestHandler<GetAllProductsQuery, PagedResponse<IEnumerable<GetAllProductsViewModel>>>
+{
+    private readonly IProductRepositoryAsync _productRepository;
+    private readonly IMapper _mapper;
+    public GetAllProductsQueryHandler(IProductRepositoryAsync productRepository, IMapper mapper)
+    {
+        _productRepository = productRepository;
+        _mapper = mapper;
+    }
+
+    public async Task<PagedResponse<IEnumerable<GetAllProductsViewModel>>> Handle(GetAllProductsQuery request, CancellationToken cancellationToken)
+    {
+        GetAllProductsParameter validFilter = _mapper.Map<GetAllProductsParameter>(request);
+        IReadOnlyList<Domain.Entities.Product> product = await _productRepository.GetPagedReponseAsync(validFilter.PageNumber, validFilter.PageSize);
+        IEnumerable<GetAllProductsViewModel> productViewModel = _mapper.Map<IEnumerable<GetAllProductsViewModel>>(product);
+        return new PagedResponse<IEnumerable<GetAllProductsViewModel>>(productViewModel, validFilter.PageNumber, validFilter.PageSize);
     }
 }
