@@ -1,15 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Application.Features.Products.Queries.GetAllProducts;
 
-namespace Application.Features.Products.Queries.GetAllProducts
+public class GetAllProductsViewModel
 {
-    public class GetAllProductsViewModel
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string Barcode { get; set; }
-        public string Description { get; set; }
-        public decimal Rate { get; set; }
-    }
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public string Barcode { get; set; }
+    public string Description { get; set; }
+    public decimal Rate { get; set; }
 }
