@@ -1,22 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Application.Parameters;
 
-namespace Application.Filters
+public class RequestParameter
 {
-    public class RequestParameter
+    public int PageNumber { get; set; }
+    public int PageSize { get; set; }
+    public RequestParameter()
     {
-        public int PageNumber { get; set; }
-        public int PageSize { get; set; }
-        public RequestParameter()
-        {
-            this.PageNumber = 1;
-            this.PageSize = 10;
-        }
-        public RequestParameter(int pageNumber, int pageSize)
-        {
-            this.PageNumber = pageNumber < 1 ? 1 : pageNumber;
-            this.PageSize = pageSize > 10 ? 10 : pageSize;
-        }
+        PageNumber = 1;
+        PageSize = 10;
+    }
+    public RequestParameter(int pageNumber, int pageSize)
+    {
+        PageNumber = pageNumber < 1 ? 1 : pageNumber;
+        PageSize = pageSize > 10 ? 10 : pageSize;
     }
 }
