@@ -1,22 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Application.Wrappers;
 
-namespace Application.Wrappers
+public class PagedResponse<T> : Response<T>
 {
-    public class PagedResponse<T> : Response<T>
-    {
-        public int PageNumber { get; set; }
-        public int PageSize { get; set; }
+    public int PageNumber { get; set; }
+    public int PageSize { get; set; }
 
-        public PagedResponse(T data, int pageNumber, int pageSize)
-        {
-            this.PageNumber = pageNumber;
-            this.PageSize = pageSize;
-            this.Data = data;
-            this.Message = null;
-            this.Succeeded = true;
-            this.Errors = null;
-        }
+    public PagedResponse(T data, int pageNumber, int pageSize)
+    {
+        this.PageNumber = pageNumber;
+        this.PageSize = pageSize;
+        this.Data = data;
+        this.Message = null;
+        this.Succeeded = true;
+        this.Errors = null;
     }
 }
