@@ -1,16 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Domain.Settings;
 
-namespace Domain.Settings
+public class MailSettings
 {
-    public class MailSettings
-    {
-        public string EmailFrom { get; set; }
-        public string SmtpHost { get; set; }
-        public int SmtpPort { get; set; }
-        public string SmtpUser { get; set; }
-        public string SmtpPass { get; set; }
-        public string DisplayName { get; set; }
-    }
+    public string EmailFrom { get; set; }
+    public string SmtpHost { get; set; }
+    public int SmtpPort { get; set; }
+    public string SmtpUser { get; set; }
+    public string SmtpPass { get; set; }
+    public string DisplayName { get; set; }
 }
