@@ -1,38 +1,37 @@
 ﻿using Application.Enums;
+
 using Infrastructure.Identity.Models;
+
 using Microsoft.AspNetCore.Identity;
-using System;
-using System.Collections.Generic;
+
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace Infrastructure.Identity.Seeds
-{
-    public static class DefaultBasicUser
-    {
-        public static async Task SeedAsync(UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
-        {
-            //Seed Default User
-            var defaultUser = new ApplicationUser
-            {
-                UserName = "basicuser",
-                Email = "basicuser@gmail.com",
-                FirstName = "John",
-                LastName = "Doe",
-                EmailConfirmed = true,
-                PhoneNumberConfirmed = true
-            };
-            if (userManager.Users.All(u => u.Id != defaultUser.Id))
-            {
-                var user = await userManager.FindByEmailAsync(defaultUser.Email);
-                if (user == null)
-                {
-                    await userManager.CreateAsync(defaultUser, "123Pa$$word!");
-                    await userManager.AddToRoleAsync(defaultUser, Roles.Basic.ToString());
-                }
+namespace Infrastructure.Identity.Seeds;
 
+public static class DefaultBasicUser
+{
+    public static async Task SeedAsync(UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
+    {
+        //Seed Default User
+        ApplicationUser defaultUser = new()
+        {
+            UserName = "basicuser",
+            Email = "basicuser@gmail.com",
+            FirstName = "John",
+            LastName = "Doe",
+            EmailConfirmed = true,
+            PhoneNumberConfirmed = true
+        };
+        if (userManager.Users.All(u => u.Id != defaultUser.Id))
+        {
+            ApplicationUser user = await userManager.FindByEmailAsync(defaultUser.Email);
+            if (user == null)
+            {
+                _ = await userManager.CreateAsync(defaultUser, "123Pa$$word!");
+                _ = await userManager.AddToRoleAsync(defaultUser, Roles.Basic.ToString());
             }
+
         }
     }
 }
