@@ -1,11 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace WebApi.Models;
 
-namespace WebApi.Models
+public class Metadata
 {
-    public class Metadata
-    {
-    }
 }
